Add files via upload
Massive overhaul working on to fix the script. This is part 1 of 2. This is to fix all dependencies needed for Linux OS's, refactor first script into segmented files for easier reading and editing individual functions, AND added a WebGUI for the script builder (Based off of my Report Builder Project's WebGUI). I also have added some additional fixes to the Orpheus_tts.py.